I've refactored the Docker setup, scripts, and documentation to get things ready for an alpha release.
Here are the key changes I made:

1.  **Script Refactoring:**
    *   I broke down `backend/app/setup_scripts.py` into smaller, more focused files:
        *   `backend/setup.py` (for Python setup utility)
        *   `scripts/start.sh`, `scripts/backup.sh`, `scripts/restore.sh`
        *   `backend/app/config.py`, `backend/app/logging_config.py`
    *   I also updated the paths and logic within these scripts so they can function independently.

2.  **Dockerfile Enhancements:**
    *   I implemented multi-stage builds in the production `Dockerfile` to create smaller image sizes.
    *   I configured both `Dockerfile` and `Dockerfile.dev` to run as a non-root user for better security.
    *   I added a `HEALTHCHECK` to `Dockerfile.dev`.

3.  **Docker Compose Improvements:**
    *   I ensured that environment variables, especially secrets, are now sourced from the `.env` file.
    *   I removed some redundant configurations, like `DATABASE_URL` in `docker-compose.yml` and the `RELOAD=true` environment variable in `docker-compose.dev.yml`.

4.  **Critical Backup/Restore Script Fixes:**
    *   I modified `scripts/backup.sh` and `scripts/restore.sh` to correctly use Docker commands for backing up and restoring the `pulseway_data` named volume. This fixes a major issue where they previously only worked with local file paths.

5.  **Comprehensive README Updates:**
    *   I updated your documentation to reflect all the script and Docker changes.
    *   I added a new "Utility Scripts" section.
    *   I significantly clarified the frontend serving status.
    *   I provided detailed explanations on database initialization and persistence, making sure to distinguish between local setups (using `backend/setup.py`) and Dockerized scenarios (which use an application-managed named volume).
    *   I updated the descriptions for the backup and restore scripts to reflect their new Docker-aware functionality.

These changes should provide a more robust, secure, and well-documented foundation for your application, getting it ready for an alpha release. Next, I planned to focus on final alpha release preparations, like versioning, pinning dependencies, and general cleanup.
