+++ Dockerfile
@@ -1,32 +1,54 @@
+# Stage 1: Builder
+FROM python:3.11-slim AS builder
+
+# Set environment variables for builder stage
+ENV PYTHONDONTWRITEBYTECODE=1 \
+    PYTHONUNBUFFERED=1
+
+# Install build dependencies
+RUN apt-get update && apt-get install -y gcc && rm -rf /var/lib/apt/lists/*
+
+# Set work directory for builder
+WORKDIR /build
+
+# Copy requirements file
+COPY requirements.txt .
+
+# Install Python dependencies into a specific prefix
+RUN pip install --no-cache-dir --prefix="/install" -r requirements.txt
+
+# Stage 2: Final image
 FROM python:3.11-slim
 
 # Set environment variables
 ENV PYTHONDONTWRITEBYTECODE=1 \
     PYTHONUNBUFFERED=1 \
-    PYTHONPATH=/app
+    PATH="/install/bin:$PATH" \
+    PYTHONPATH="/install/lib/python3.11/site-packages"
+
+# Create a non-root user and group
+RUN groupadd -r appgroup && useradd --no-log-init -r -g appgroup appuser
 
 # Set work directory
 WORKDIR /app
 
-# Install system dependencies
-RUN apt-get update && apt-get install -y \
-    gcc \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements first for better caching
-COPY requirements.txt .
-
-# Install Python dependencies
-RUN pip install --no-cache-dir -r requirements.txt
+# Copy installed dependencies from builder stage
+COPY --from=builder /install /usr/local
 
 # Copy application code
 COPY ./backend/ /app/
 
-# Create directory for SQLite database
-RUN mkdir -p /app/data
+# Create and set permissions for data directory (before switching user)
+# /app/data is also a volume mount point, permissions might be managed by Docker at runtime.
+# However, ensuring the directory exists with correct ownership is good practice.
+RUN mkdir -p /app/data && chown -R appuser:appgroup /app/data
+RUN chown -R appuser:appgroup /app
 
 # Expose port
 EXPOSE 8000
+
+# Switch to non-root user
+USER appuser
 
 # Health check
 HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
@@ -34,39 +56,3 @@ HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
 
 # Run the application
 CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
-
-# Dockerfile.dev - Development Dockerfile
-FROM python:3.11-slim
-
-ENV PYTHONDONTWRITEBYTECODE=1 \
-    PYTHONUNBUFFERED=1 \
-    PYTHONPATH=/app
-
-WORKDIR /app
-
-# Install system dependencies including curl for health checks
-RUN apt-get update && apt-get install -y \
-    gcc \
-    curl \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements and install dependencies
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Install development dependencies
-RUN pip install --no-cache-dir \
-    pytest \
-    pytest-asyncio \
-    black \
-    flake8 \
-    mypy
-
-# Create data directory
-RUN mkdir -p /app/data /app/logs
-
-# Expose port
-EXPOSE 8000
-
-# Default command (can be overridden in docker-compose.dev.yml)
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000", "--reload"]
