+++ backend/app/logging_config.py
@@ -1,0 +1,56 @@
+"""
+Logging configuration for Pulseway Backend
+"""
+
+import logging
+import logging.handlers
+import os
+from pathlib import Path
+
+def setup_logging(log_level: str = "INFO"):
+    """Setup application logging"""
+
+    # Create logs directory
+    log_dir = Path("logs")
+    log_dir.mkdir(exist_ok=True)
+
+    # Configure root logger
+    logger = logging.getLogger()
+    logger.setLevel(getattr(logging, log_level.upper()))
+
+    # Clear existing handlers
+    logger.handlers.clear()
+
+    # Console handler
+    console_handler = logging.StreamHandler()
+    console_handler.setLevel(logging.INFO)
+    console_format = logging.Formatter(
+        '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+    )
+    console_handler.setFormatter(console_format)
+    logger.addHandler(console_handler)
+
+    # File handler with rotation
+    file_handler = logging.handlers.RotatingFileHandler(
+        log_dir / "pulseway_backend.log",
+        maxBytes=10*1024*1024,  # 10MB
+        backupCount=5
+    )
+    file_handler.setLevel(logging.DEBUG)
+    file_format = logging.Formatter(
+        '%(asctime)s - %(name)s - %(levelname)s - %(funcName)s:%(lineno)d - %(message)s'
+    )
+    file_handler.setFormatter(file_format)
+    logger.addHandler(file_handler)
+
+    # Error file handler
+    error_handler = logging.handlers.RotatingFileHandler(
+        log_dir / "pulseway_errors.log",
+        maxBytes=5*1024*1024,  # 5MB
+        backupCount=3
+    )
+    error_handler.setLevel(logging.ERROR)
+    error_handler.setFormatter(file_format)
+    logger.addHandler(error_handler)
+
+    return logger
